Genome Analyzer/API/requirements.txt
adal==1.2.7
aniso8601==9.0.1
astroid==2.4.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
bio==0.7.4
biopython==1.79
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.3
contextlib2==0.6.0.post1
cryptography==3.4.7
cycler==0.10.0
dnspython==2.1.0
docker==4.4.4
email-validator==1.1.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.3
Flask-Login==0.5.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
isodate==0.6.0
isort==4.3.21
itsdangerous==2.0.1
jeepney==0.7.1
Jinja2==3.0.1
jmespath==0.10.0
jsonpickle==2.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
matplotlib==3.4.1
mccabe==0.6.1
msrest==0.6.21
msrestazure==0.6.4
ndg-httpsclient==0.5.1
numpy==1.20.2
oauthlib==3.1.1
pathspec==0.9.0
Pillow==8.2.0
plac==1.3.3
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
pylint==2.5.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.6
SecretStorage==3.3.1
six==1.14.0
SQLAlchemy==1.4.22
toml==0.10.0
tqdm==4.62.2
urllib3==1.26.5
websocket-client==1.2.1
Werkzeug==2.0.1
wrapt==1.12.1
WTForms==2.3.3